Check api in Home, allow null values.
